refactor methods into module and add railtie
