+++ test/dummy/app/models/vhost.rb
@@ -1,5 +1,5 @@
 class Vhost < ActiveRecord::Base
-  include CommittedChanges
+  # include CommittedChanges
   validates_presence_of :domains
   after_commit :reconfigure_server, :if => :domains_changed_and_committed?
